add : New interfaces for swap decision and solving logic
What has been created :
An 'ISubSolver' class to take care of the Follower's heuristics.
An 'IFollowerSwapRule' class to represent a decision rule when checking if a swap is possible. It will be used in the 'ISubSolver' class.
An 'ILeaderSelectRule' class to represent a selection rule when the Leader must choose a subset oj jobs. It will be used in 'ISolver' class.