add : New solver structure
Structure of classes changed.
What was added :
A 'FLateToEarly' class which represent a Follower's decicion rule when checking possible swaps
A 'FSolver' class which implements the 'ISubSolver' class by applying a heuristic to solve the Follower's lexicographical problem
A 'LSolver' class which implements the 'ISolver' class by applying a heuristic to solve the Leader's problem.

What was updated :
New getter for 'Solution' class to get a Machine as a const object.
Commented 'IFollowerSwapRule' methods
The 'startTimeOfJob' method now returns a const float instead of a float.