update : Gathered classes' method into one class
'LLPTRule' class is renamed 'LSortRule' and contains the definition of the 'selectJobFrom' method of the 'LLPTRule' class and 'LWPTRule' class
'LWPTRule' class has been removed

'LSortRule' now is initialized with a value from the 'sortRule' namespace and will call the corresponding method's definition when its 'selectJobsFrom' method is called.

This structural change was done to reduce the number of .h files in the project.
If 'LSortRule' becomes filled with too many methods, a child class will be added to contains newer methods and so on.