update : Added SwapOperation structure
A swapOperation struct has been added in the class 'IFollowerSwapRule'.
It contains the following data :
machine1 - the first machine to perform the swap to
machine2 - the second machine to perform the swap to
bloc - the position of the machines' job
gain - How much will this operation improve the solution (higher -> better)

The 'FSolver' class has been updated to handle the structure returned by the 'IFollowerSwapRule' methods

The 'bestSwapV' method in 'FLateToEarly' now returns the 'SwapOperation' with the highest gain.