add : New Selection Rule for the Leader (SPT_EDD_CONST_LATENESS)
This rule will sort jobs by increasing processing time and earliest due date.
Assuming we only have one machine with a speed equal to 1, each job's lateness is calculated.
The jobs with a non positive lateness are selected.
If there is still an amount of jobs to select, those with the smallest positive lateness are selected until we reach the number of jobs to select.