add : Time resolution, adding rules to solvers, Heuristic's name and description
Added methods to return the heuristic's name and description of a 'ISolver' child class such as 'LSolver'.
Added an adding feature for a 'ISolver' 's and 'ISubSolver' 's list of rules.
Added an attribute in the 'ISolver' class to store the time spent to solve an instance. This attribute is initialized at 0 and updated when the method 'solve()' is called.